feat: unroll container type en-/decoding
Instead of relying on generic implementations for the Rust generated
code that en- or decodes the data, these are defined as closures.

The main benefit is that it allows optimized calls that would not be
possible through the trait system (for example a fast encoding of
`Vec<u8>` instead of the regular `Vec<T>` encoding that would write the
bytes one-by-one).