refactor(lsp): replace tower-lsp with lsp-server
The `tower-lsp` crate appears to be somewhat unmaintained in the recent
months and the `lsp-server` crate is made by the people behind
rust-analyzer.

In addition, the LSP server doesn't need a full async runtime as LSP
communication is only with a single client and done in serial.

This change cuts down on several dependencies, reducing both build time
and binary size.