feat(doc): implement a HTML documentation generator
A new generator that doesn't create implementations for a language, but
instead creates HTML documentation for a schema.

This is the first building block to provide a new CLI command that can
generate the documentation for any schema, similar to how rustdoc can
generate local documentation for Rust files or projects.