feat(compiler): introduce a simplified schema for easier codegen
The parser's data structures can be a bit hard to use in the code
generator, and some upcoming features would need to be rebuilt in every
code generator.

By introducing a simplified version of the schema data structures,
common conversion logic can be moved into the compiler and overall make
the data structures more convenient to use.