feat: allow to omit IDs and derive them instead
Similar to Rust basic enums, the identifier for each variant doesn't
have to be specified explicitly. Instead it can be derived by continuing
to count upwards from the last explicit value.

Therefore, now the field IDs as well as enum variant IDs can be omitted
and are derived by the compiler. They can be mixed as well to create
gaps in the ID range.