feat(parser): add `simd` feature to the parser
The `winnow` crate has a `simd` feature which enables some extra
performance improvements. This might limit possible target platforms for
compilations so the flag is exposed as optional feature on the parser
crate instead.

All binary crates will use this flag, but potential users of the parser
crate can omit the feature if it causes issues.