feat: change type alias to plain alias (including the keyword)
Type aliasing has been named inconsistently. The keyword is `type`,
often it's called type alias in the code, and sometimes just alias.

As types are the only thing that can be aliased in a schema, it is now
simply named `alias` everywhere, with the keyword changed from `type` to
`alias` as well.