feat: track keywords, punctuation and delimiters
In addition to the more valuable parts of a schema that describe the
data structures, now the remaining pieces that make up a schema are
tracked as well.

That means characters like commas, semicolons, open/close braces and
others are collected so their exact location in the source file can be
used by the LSP and other components.

A few places still don't track this information as they need some
further refactoring first, but should be doing so in the following
commits.