refactor: introduce a container for punctuation
Many elements in Mabo schema files use some form of punctuation to
separate from another. This common logic could be centralized in a
custom container that keeps track of punctuation spans and takes care of
formatting logic.

This adds missing tracking of token spans that haven't be tracked before
yet.